refactor(workflow): extract grouped rollout wrapper
